(maint) Teach Vanagon how to handle pre and post remove actions
Prior to this commit, Vanagon only knew how to handle adding preinstall
and postinstall packaging actions. This commit adds the ability for
Vanagon to parse pre and post removal packaging actions. It also reworks
how pre and post install actions are handled to make them more powerful.
They can now define how to handle actions that should ony be run on
install or upgrade or both. The same is true for pre and post removal
actions, where we can handle actions that should only be run on upgrade
or removal or both.