(CPR-235) Teach Vanagon how to add packaging scriptlets to debian packaging
Before this commit, Vanagon was unable to add in any packaging scripts
to the debian packaging. This commit adds in that ability and also
delivers a preinst file to packaging. This commit also tells Debian how
to run certain scripts based on defined characteristics, like if a
script should be run on upgrade, install, or both.