(CPR-234) Teach Vanagon how to add packaging scriptlets to rpm packaging
Prior to this commit, rpm packaging only new how to handle pre and post
install scripts, and had no ability to handle special cases, like if we
wanted to only run a script on install and not upgrade. This commit adds
the ability to handle pre and post removal scripts, and adds the ability
to handle special instructions, even in the %posttrans section.