NOT WORKING
current problem I was trying to solve: if we're using a precompiled
artifact, packaging works fine on platforms that don't have any extra
files, but adding in extra init / config / etc files were causing
failures. Need to figure out how to pick and choose which ones end up in
which package

Adding all the files to the tarball resulted in them all making it into
all packages.

Was testing with builds of pe-nginx on the 'generics' branch of
puppet-enterprise-vanagon.